requirements.txt
absl-py==1.0.0
bagnets @ git+https://github.com/wielandbrendel/bag-of-local-features-models.git@8493fb442c9b9ccfeaa970c950ca67832017f072
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
contourpy==1.0.5
cycler==0.11.0
fonttools==4.38.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
importlib-metadata==4.11.3
joblib==1.1.0
kiwisolver==1.4.4
Markdown==3.3.6
matplotlib==3.6.0
multi-level-split @ git+https://github.com/lmkoch/multi-level-split@638ec077fe7e5494ef5df87f34828df300704b2f
munch==2.5.0
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.0
pandas==1.4.2
Pillow==9.1.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.12.1
six==1.16.0
sklearn==0.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.50.2
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
grad-cam==1.3.9
